Disable additional metadata collection
As suggested in this issue [1]. This should avoid warnings like this
from appearing in e.g. `cdk diff` output:

    [Warning at /edge-lambda-stack-c88816ed891470b4969caff927b6a3c6a2c676b4fa/hannahsmithsonOrgRedirect/ServiceRole]
    Failed to add construct metadata for node [ServiceRole].
    Reason: ValidationError: The result of fromAwsManagedPolicyName can not be used in this API
    [ack: @aws-cdk/core:addConstructMetadataFailed]

It seems as if the EdgeFunction construct is generating a role with an
AWS-managed policy from which AWS CDK cannot extract metadata.
Apparently this doesn't indicate a problem so it should be safe to
disable the metadata collection.

[1]: https://github.com/aws/aws-cdk/issues/36078